+++ src/Model/TimeCalculator.java
@@ -30,11 +30,6 @@ public class TimeCalculator {
 		endTime = StringToLocal(end);
 	}
 
-	// nowTime 을 String으로 입력받고, local변수로 초기화 해주는 생성자
-	// strattime과 endtime 을 string으로 받고 local변수로 변환해주는 메소드
-	// 현재 local 변수에 now, start, end, 를 가지고 포함되어 있는지 아닌지 t/f로 반환하는 메소드
-	// 남
-
 	public LocalTime StringToLocal(String time){
 		String[] arr = time.split(":");
 		int[] arrInt = new int[arr.length];
@@ -45,11 +40,9 @@ public class TimeCalculator {
 	
 	public void compareTime() {
 
-		//사용중일 때
+		//사용중일 때 -> (시작시간하고 끝시간 사이인경우 || 현재시각이랑 시작시간이 같은 경우)
 		if (nowTime.isAfter(startTime) && nowTime.isBefore(endTime) || nowTime.equals(startTime)) {
-			//사용중이면 연속적으로 강의가 있는지 Left Time 에 계산할 필요가 있다.
-			//근본적으로 endTime 값을 수정해서 가져오는게 현실적이지 않은가?
-			result =  "Classroom is now using, left time for empty is : " + nowTime.until(endTime, ChronoUnit.MINUTES) + " minutes";
+			result =  "현재 강의실이 사용중입니다. " + nowTime.until(endTime, ChronoUnit.MINUTES) + "분 후에 강의가 마칩니다.";
 			LeftTime = nowTime.until(endTime, ChronoUnit.MINUTES);
 			isEmpty = false;
 		}
@@ -58,7 +51,7 @@ public class TimeCalculator {
 		else if (nowTime.isBefore(startTime)) {
 			if(LeftTime > nowTime.until(startTime, ChronoUnit.MINUTES)){
 				LeftTime = nowTime.until(startTime, ChronoUnit.MINUTES);
-				result = "Classroom is now empty, left time for next class is : " + LeftTime + " minutes";
+				result = "현재 강의실이 비어있습니다. 다음 수업까지는 " + LeftTime + "분 남았습니다.";
 			}
 		}
