+++ src/Test/ModelTest.java
@@ -1,6 +1,7 @@
 package Test;
 
-import Model.ITBuild;
+import Model.ParsingClass;
+import Model.ReadData;
 
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
@@ -8,22 +9,25 @@ import java.util.ArrayList;
 public class ModelTest {
 
     public static void main(String[] args) throws FileNotFoundException {
-        ITBuild itBuild = new ITBuild();
-        itBuild.setArrayLists(); //파일에서 데이터 읽어옴
-        for (ArrayList<String> arrayList : itBuild.getArrayLists()) {
+        ReadData readData = new ReadData();
+        readData.setArrayLists(); //파일에서 데이터를 읽어온다.
+
+        for (ArrayList<String> arrayList : readData.getArrayLists()) {
             System.out.println("arrayList = " + arrayList);
         }
-        itBuild.setCinfos(); // 읽어온 데이터를 Cinfos에 저장함
-        itBuild.showAllClassInfo(); // Cinfos 를 모두 출력
-//
-        itBuild.setClassByRoom(); // 호실 마다 시간표를 정리
-        itBuild.printClassByRoom(); // 호실 배정된 시간표를 출력
 
+        ParsingClass parsingClass = new ParsingClass();
+
+        parsingClass.settingClassInfos(); // 읽어온 데이터를 Cinfos에 저장함
+        parsingClass.showAllClassInfo(); // Cinfos 를 모두 출력
+//
+        parsingClass.setClassByRoom(); // 호실 마다 시간표를 정리
+        parsingClass.printClassByAllRooms(); // 호실 배정된 시간표를 출력
 
         // 특정시간을 매개변수로 넣고 비어있는지 차 있는지 확인한다.
-        for (String room : itBuild.getClassByRoom().keySet()) {
+        for (String room : parsingClass.getClassByRoom().keySet()) {
             System.out.println("room : " + room);
-            System.out.println(itBuild.IsEmptyRoom(room, "월", "11:11"));
+            System.out.println(parsingClass.IsEmptyRoom(room, "월", "11:11"));
         }
